Extract Dynamic Workflow spec validation out of store.py into validation.py
Mechanical move of validate_workflow_spec, validate_workflow_input,
workflow_runtime_seconds, DynamicWorkflowError, and all private
_validate_* helpers plus their constants into
src/mindroom/dynamic_workflows/validation.py. The helpers shared with
the store (ID_RE, validate_id, object_mapping) become public module
symbols. store.py keeps persistence and report rendering and imports
the validation entry points it still uses; all callers now import the
validation surface from validation.py directly (no re-export aliases).
